Provider specific locales: provider specific locales are identified eg for OpenStack: en_OS Locale is set in set_locale method of application controller.
This patchset introduces provider specific translations for winged monkey, so the UI will present the user with the terms/naming of the selected provider. (eg. with Mock, the user will see "My Applications", in OpenStack user will see "My Server Instances").

This functionality is implemented using gettext locales, so when user selects provider on login page, the I18n.locale is set to en_{provider.locale_id}. Locale_id is set in providers.yml. It is not possible to use provider type, because locales have to be in aa_BB format. (eg. locale_id for OpenStack provider is "OS").

Translating:
- we have default locales that carry default translations (used when no provider is selected), now we have only config/gettext_locales/en/
- for each provider we have to create new locale (eg. config/gettext_locales/en_OS) which will translate provider-specific terms

Note that all these files have to be the same, only msgstr's will differ. It is not possible to have only a subset of strings in provider_specific locale .po file.

Updated rubyparser gem